muphasa_invoice/requirements.txt
eml2png
pypdf2
pytesseract
cohere
mysql-connector-python
python-dateutil
# pytest